[nemo-qml-plugin-calendar] Use notebook color if specified
Previously if no setting for the notebook color was defined, the
notebook would be colored randomly.  This commit ensures that if
no explicit setting for the notebook color was defined, it should
instead attempt to use the color specified by the notebook, and
only if that is empty, to fall back to the randomly generated
color.