[nemo-qml-plugin-calendar] Enablers for DB operations in thread
The NemoCalendarManager class controls a thread implemented by
class NemoCalendarWorker which runs all DB operations.

NemoCalendarManager provides an API through which calendar data
is accessed. Most of the functions in the API are asynchronous,
results are reported through signals. The two exceptions are
the synchronous functions NemoCalendarManager::createEvent(),
and NemoCalendarManager::convertEventToVCalendarSync().

NemoCalendarManager keeps a cache of loaded events and occurrences,
and consults the cache when it gets requests for calendar data.
In addition, NemoCalendarWorker prepares the data structures used
by NemoCalendarManager to keep calendar data processing in the GUI
thread to a minimum.

NemoCalendarManager and NemoCalendarWorker load data from DB into
memory by time periods. NemoCalendarManager stores the loaded
periods so that it knows what data is currently cached, and what
data needs to be reloaded if the DB is changed by some other
software entity. The periods are automatically managed by
NemoCalendarManager.