[nemo-qml-plugin-calendar] Emit updated() on AgendaModel refresh
The updated() signal is emitted when AgendaModel model data has
been updated either as a response to changes in startDate and/or
endDate, or changes in the data base.

The updated() signal does not imply that the AgendaModel has been
changed, only that the contents of the model are up to date.