Add property eventDisplayTime CalendarEventsModel
The property controls how the model expiry date is calculated. If
eventDisplayTime <= 0, then the expiry date is the next end time
of events included in the model.

If eventDisplayTime > 0, then the event end times are calculated
by adding eventDisplayTime to the start times of events. The
expiryDate property will then reflect the next start time +
eventDisplayTime of events included in the model.